feat: view as user component
